+++ Firmware/OTPManagerFirmware/crc8.h
@@ -3,6 +3,6 @@
 
 #include <Arduino.h>
 
-uint8_t CRC8(const uint8_t *source, uint8_t len);
+uint8_t CRC8(const uint8_t *source, size_t len);
 
 #endif
